# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

/**
 * A slim collection reference embedded in product responses
 */
export interface ProductCollection {
  id: number;
  name: string;
  slug: string;
}

export interface Collection {
  id: number;
  name: string;
  slug: string;
  /** Filter-panel section: 'category' | 'collection' | 'holiday' | null */
  grp?: string | null;
  sortOrder: number;
  productCount: number;
  availableCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CollectionInput {
  /** @minLength 1 */
  name: string;
  slug?: string;
  grp?: string | null;
  sortOrder?: number;
}

export interface CollectionUpdate {
  /** @minLength 1 */
  name?: string;
  slug?: string;
  grp?: string | null;
  sortOrder?: number;
}

export interface Product {
  id: number;
  name: string;
  collections: ProductCollection[];
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  sku: string | null;
  /** @nullable */
  size: string | null;
  /** @nullable */
  description: string | null;
  available: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minItems 1 */
  collectionIds: number[];
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  sku?: string | null;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  description?: string | null;
  available?: boolean;
  sortOrder?: number;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minItems 1 */
  collectionIds?: number[];
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  sku?: string | null;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  description?: string | null;
  available?: boolean;
  sortOrder?: number;
}

/**
 * A single product row for bulk import. Collections are matched by name and auto-created if new.
 */
export interface BulkProductInput {
  /** @minLength 1 */
  name: string;
  /** @minItems 1 */
  collectionNames: string[];
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  sku?: string | null;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  description?: string | null;
  available?: boolean;
  sortOrder?: number;
}

export interface BulkCreateProductsInput {
  /** @maxItems 2000 */
  products: BulkProductInput[];
}

export interface BulkCreateProductsResult {
  created: number;
  duplicates: number;
  failed: number;
  errors?: string[];
}

export interface BulkDeleteProductsInput {
  /** @minItems 1 */
  ids: number[];
}

export interface BulkDeleteProductsResult {
  deleted: number;
}

export interface AdminLoginInput {
  /** @minLength 1 */
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
}

export type ApplicationStatus = typeof ApplicationStatus[keyof typeof ApplicationStatus];


export const ApplicationStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Application {
  id: number;
  businessName: string;
  contactName: string;
  email: string;
  phone: string;
  businessType: string;
  /** @nullable */
  monthlyVolume: string | null;
  /** @nullable */
  notes: string | null;
  status: ApplicationStatus;
  createdAt: string;
  updatedAt: string;
}

export interface ApplicationInput {
  /** @minLength 1 */
  businessName: string;
  /** @minLength 1 */
  contactName: string;
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  businessType: string;
  monthlyVolume?: string | null;
  notes?: string | null;
}

export type ApplicationUpdateStatus = typeof ApplicationUpdateStatus[keyof typeof ApplicationUpdateStatus];


export const ApplicationUpdateStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface ApplicationUpdate {
  status?: ApplicationUpdateStatus;
  notes?: string | null;
}

export interface UploadRequestInput {
  name: string;
  size: number;
  contentType: string;
}

export interface PresignedUpload {
  uploadURL: string;
  objectPath: string;
}

export interface BrochureAuthInput {
  /** @minLength 1 */
  password: string;
}

export interface BrochureSession {
  authenticated: boolean;
}

export interface Brochure {
  id: number;
  title: string;
  season: string;
  objectPath: string;
  fileName: string;
  createdAt: string;
}

export interface BrochureInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  season: string;
  /** @minLength 1 */
  objectPath: string;
  /** @minLength 1 */
  fileName: string;
}

export interface PriceList {
  id: number;
  title: string;
  period: string;
  objectPath: string;
  fileName: string;
  createdAt: string;
}

export interface PriceListInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  period: string;
  /** @minLength 1 */
  objectPath: string;
  /** @minLength 1 */
  fileName: string;
}

export type ListCollectionsParams = {
availableOnly?: boolean;
};

export type ListProductsParams = {
collectionId?: number;
/**
 * Filter by exact size value
 */
size?: string;
availableOnly?: boolean;
};