# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /brochures/auth:
    post:
      operationId: brochureAuth
      tags: [brochures]
      summary: Authenticate as a wholesale buyer to access brochures
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrochureAuthInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrochureSession"
        "401":
          description: Incorrect password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: getBrochureSession
      tags: [brochures]
      summary: Check whether the current session has brochure access
      responses:
        "200":
          description: Session status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrochureSession"
  /brochures:
    get:
      operationId: listBrochures
      tags: [brochures]
      summary: List all brochures (requires brochure or admin session)
      responses:
        "200":
          description: List of brochures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brochure"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /brochures/{id}/download:
    get:
      operationId: downloadBrochure
      tags: [brochures]
      summary: Stream a brochure PDF (requires portal or admin session)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF file stream
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/brochures:
    post:
      operationId: createBrochure
      tags: [brochures]
      summary: Register a brochure record after client-side upload (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrochureInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brochure"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/brochures/{id}:
    delete:
      operationId: deleteBrochure
      tags: [brochures]
      summary: Delete a brochure record (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /price-lists:
    get:
      operationId: listPriceLists
      tags: [priceLists]
      summary: List all price lists (requires portal or admin session)
      responses:
        "200":
          description: List of price lists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceList"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /price-lists/{id}/download:
    get:
      operationId: downloadPriceList
      tags: [priceLists]
      summary: Stream a price list PDF (requires portal or admin session)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF file stream
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/price-lists:
    post:
      operationId: createPriceList
      tags: [priceLists]
      summary: Register a price list record after client-side upload (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceListInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceList"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/price-lists/{id}:
    delete:
      operationId: deletePriceList
      tags: [priceLists]
      summary: Delete a price list record (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    BrochureAuthInput:
      type: object
      required:
        - password
      properties:
        password:
          type: string
          minLength: 1
    BrochureSession:
      type: object
      required:
        - authenticated
      properties:
        authenticated:
          type: boolean
    Brochure:
      type: object
      required:
        - id
        - title
        - season
        - objectPath
        - fileName
        - createdAt
      properties:
        id:
          type: integer
        title:
          type: string
        season:
          type: string
        objectPath:
          type: string
        fileName:
          type: string
        createdAt:
          type: string
          format: date-time
    BrochureInput:
      type: object
      required:
        - title
        - season
        - objectPath
        - fileName
      properties:
        title:
          type: string
          minLength: 1
        season:
          type: string
          minLength: 1
        objectPath:
          type: string
          minLength: 1
        fileName:
          type: string
          minLength: 1
    PriceList:
      type: object
      required:
        - id
        - title
        - period
        - objectPath
        - fileName
        - createdAt
      properties:
        id:
          type: integer
        title:
          type: string
        period:
          type: string
        objectPath:
          type: string
        fileName:
          type: string
        createdAt:
          type: string
          format: date-time
    PriceListInput:
      type: object
      required:
        - title
        - period
        - objectPath
        - fileName
      properties:
        title:
          type: string
          minLength: 1
        period:
          type: string
          minLength: 1
        objectPath:
          type: string
          minLength: 1
        fileName:
          type: string
          minLength: 1