# lib/api-zod/src/generated/types/priceList.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PriceList {
  id: number;
  title: string;
  period: string;
  objectPath: string;
  fileName: string;
  createdAt: Date;
}