# lib/api-zod/src/generated/types/priceListInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PriceListInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  period: string;
  /** @minLength 1 */
  objectPath: string;
  /** @minLength 1 */
  fileName: string;
}